Archive existing ULS snapshot and point to WFA snapshot
Removes the existing ULS snapshot and adds a README file pointing users to the current WFA snapshot. This link is specifically to the current version and will not be affected if a later version is uploaded. A link to the removed WF snapshot is included for archival purposes, as are instructions on how to be notified of any changes to the WFA ULS Snapshot FCC weekly subdirectory.