Correct the FrequencyRange validation logic
Intent was to check that low frequency is strictly less than high frequency, but implementation allowed them to be equal. This corrects the comparison to strictly less than.

Note: This requirement is not strictly stated in the SDI spec, but it resolves ambiguity regarding the interpretation of adjacent frequency ranges. For example, we choose to interpret ranges of 6020- 6050 and 6050-6070) as the frequencies from 6020 up to but not including 6050 (i.e., [6020, 6050)), and 6050 up to but not including 6070 (i.e., [6050, 6070)). The example response in the SDI spec allows for the same frequency to be a low and a high freq in different ranges simultaneously, and interpreting ranges as being inclusive of both endpoints leads to conflicting limits. As such, a single MHz frequency range (6050 up to but not including 6051) should be specified as {low: 6050, high: 6051} and not {low: 6050, high: 6050}.