Temp fix for definition of AFCS.URS.1 response mask
This modifies AFCS.URS.1_mask.json to expect response code 102 (MISSING_PARAM)
in addition to the original 103 (INVALID_VALUE). The request for this
vector is missing a device ID, so 102 is the more appropriate response code.
Wi-Fi Alliance has requested the response mask file be updated to allow
both code 102 and 103, pending a more formal resolution or update to
their SUT test vector document.