Relax HTTP status code check
SDI is ambiguous on HTTP status code for many cases. Specifically, it's
unclear what the status code should be if the response message contains
any responses that indicate a non-SUCCESS code.

Originally, this harness assumed that a 2XX HTTP code would be returned
for any response that did not meet one of the explicit criteria listed
in the SDI (missing HTTP headers, bad URL, missing SDI message).
However, it was requested that an AFC be permitted to return a 400 HTTP
status code (BAD REQUEST) alongside non-SUCCESS SDI response codes.

Because WINNF-TS-3007 V1.2.0 Section 6.3.5 states that an HTTP status
code other than 200 shall be interpreted as a failure to obtain a list
of available frequencies or channels, this harness will only permit
HTTP code 400 if all responses within a response message indicate an
error, as a combination of HTTP 400 alongside any SDI SUCCESS codes is
contradictory.

For example, if an AFC response satisfies a mask that expects any
SUCCESS codes, but the AFC response includes any non-2XX HTTP status
code, then the test result will still be UNEXPECTED. If an AFC response
satisfies a mask that expects only non-SUCCESS codes, then the harness
will accept 2XX or 400 as the HTTP status code for an EXPECTED result.