Improve robustness of list-of-dictionary conversions
Previously, the init_from_dicts function would encounter an exception
when attempting to create a dataclass list from a list of dictionaries
if any of the contained dictionaries could not be successfully converted
to dataclass form. In this instance, none of the provided dictionaries
would be converted to dataclass form, limiting the validation and tests
that could be performed on the object in question.

This commit alters init_from_dicts to catch any exceptions that occur
during dictionary conversion, leaving the original dictionary in place
in the list, but allowing other dictionaries to be converted, when
possible.

It also modifies some of the error suppression in some of the SDI
modules' post_init() functions to properly handle errors encountered
when attempting to operate on a partially converted dataclass/dictionary
list.