Change example AFC config to use a large timeout override
Per request, the example AFC config provided in this repo includes an
out-of-the-box override of 15 minutes from the default of 10 seconds.

Existing AFC config files can adopt this change by uncommenting the
timeout field (if necessary) and changing the value to 900.

This change has also been documented in the main README in src/harness.
The placeholder {commit-date} will be replaced with the date upon which
this commit is pulled into the main branch.